2_CICD/docker/docker_cicd_lab1/requirements.txt
Brotli==1.0.9
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
ConfigArgParse==1.5.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.1
geventhttpclient==2.0.8
gitdb==4.0.9
GitPython==3.1.29
greenlet==1.1.3.post0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.12.2
MarkupSafe==2.1.1
msgpack==1.0.4
psutil==5.9.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
roundrobin==0.0.4
six==1.16.0
smmap==5.0.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zope.event==4.5.0
zope.interface==5.5.0
